+++ cityscapesscripts/evaluation/evalObjectDetection3D.py
@@ -17,15 +17,11 @@ from tqdm import tqdm
 from copy import deepcopy
 import concurrent.futures
 
-from typing import (
-    Dict, 
-    List, 
-    Tuple, 
-    Union
-)
-
 from cityscapesscripts.helpers.labels import labels
 from cityscapesscripts.evaluation.objectDetectionHelpers import (
+    Box3DObject,
+    IgnoreObject,
+    EvaluationParameters,
     getFiles,
     calcIouMatrix,
     calcOverlapMatrix
@@ -70,71 +66,6 @@ def printErrorAndExit(msg):
     logger.info("========================")
     
     exit(1)
-
-class Box3DObject:
-    """Helper class storing information about a 3D-Box-instance.
-
-    Attributes:
-        box_2d_modal: modal 2d box
-        box_2d_amodal: amodal 2d box
-        center: center in 3D space
-        dims: dimensions in 3D
-        rotation: rotation of object in quaternion
-        class_name: class name in cityscapes name format
-        score: predicted score
-    """
-    def __init__(
-        self,
-        annotation: dict,
-    ) -> None:
-
-        self.box_2d_modal = annotation["2d"]["modal"]
-        self.box_2d_amodal = annotation["2d"]["amodal"]
-        self.center = annotation["3d"]["center"]
-        self.dims = annotation["3d"]["dimensions"]
-        self.rotation = annotation["3d"]["rotation"]
-        self.class_name = annotation["class_name"]
-        self.score = annotation["score"]
-
-    def getDepth(self):
-        return np.sqrt(self.center[0]**2 + self.center[2]**2).astype(int)
-
-class IgnoreObject:
-    """Helper class storing information about an ignore region.
-
-    Attributes:
-        box_2d: Coordinates of 2d-bounding box,
-    """
-    def __init__(
-        self,
-        annotation: dict,
-    ) -> None:
-
-        self.box_2d = annotation["2d"]
-
-
-class EvaluationParameters:
-    """Helper class managing the evaluation parameters
-
-    Attributes:
-        labels_to_evaluate: list of labels to evaluate
-        min_iou_to_match_mapping: min iou required to accept as TP
-        max_depth: max depth for evaluation
-        step_size: step/bin size for DDTP metrics
-    """
-
-    def __init__(
-        self,
-        labels_to_evaluate: List[str],
-        min_iou_to_match_mapping: float=0.7,
-        max_depth: int=100,
-        step_size: int=5
-    ) -> None:
-
-        self.labels_to_evaluate = labels_to_evaluate
-        self.min_iou_to_match_mapping = min_iou_to_match_mapping
-        self.max_depth = max_depth
-        self.step_size = step_size
 
 class Box3DEvaluator:
     def __init__(
@@ -935,7 +866,13 @@ def main():
     )
     predictionFolder = os.path.join(predictionPath, "box3Dpred")
 
+    ###
+    ### TMP
+    ###
     gtFolder = predictionFolder = "/lhome/ngaehle/Desktop/cs_QC_final_export_TEST/export/val/"
+    ###
+    ###
+    ###
 
     parser = argparse.ArgumentParser()
     # setup location
